feat: check permissions without breaking anything(no BR change)
