+++ crate/server/src/database/redis/instance.rs
@@ -7,7 +7,6 @@ use tracing::info;
 pub(crate) struct Redis<const WORD_LENGTH: usize> {
     pub(crate) memory: RedisMemory<Address<SERVER_ADDRESS_LENGTH>, [u8; WORD_LENGTH]>,
     pub(crate) manager: ConnectionManager,
-    // pub(crate) lock: Mutex<()>,
 }
 
 impl<const WORD_LENGTH: usize> Redis<WORD_LENGTH> {
@@ -29,10 +28,6 @@ impl<const WORD_LENGTH: usize> Redis<WORD_LENGTH> {
 
         let memory = RedisMemory::connect_with_manager(manager.clone()).await?;
 
-        Ok(Self {
-            memory,
-            manager,
-            // lock: Mutex::new(()),
-        })
+        Ok(Self { memory, manager })
     }
 }
