+++ crate/server/src/routes/datasets.rs
@@ -14,10 +14,7 @@ use crate::{
     core::FindexServer,
     database::database_traits::DatasetsTrait,
     error::result::FResult,
-    routes::{
-        check_permission,
-        error::{ResponseBytes, SuccessResponse},
-    },
+    routes::error::{ResponseBytes, SuccessResponse},
 };
 
 #[post("/datasets/{index_id}/add_entries")]
@@ -31,7 +28,9 @@ pub(crate) async fn datasets_add_entries(
 
     info!("user {user}: POST /datasets/{index_id}/add_entries");
 
-    check_permission(&user, &index_id, Permission::Write, &findex_server).await?;
+    findex_server
+        .check_permission(&user, &index_id, Permission::Write)
+        .await?;
 
     let index_id = Uuid::parse_str(&index_id)?;
     let encrypted_entries = EncryptedEntries::deserialize(&bytes)?;
@@ -66,7 +65,9 @@ pub(crate) async fn datasets_del_entries(
 
     info!("user {user}: POST /datasets/{index_id}/delete_entries");
 
-    check_permission(&user, &index_id, Permission::Write, &findex_server).await?;
+    findex_server
+        .check_permission(&user, &index_id, Permission::Write)
+        .await?;
 
     let index_id = Uuid::parse_str(&index_id)?;
     let uuids = Uuids::deserialize(&bytes)?;
@@ -101,7 +102,9 @@ pub(crate) async fn datasets_get_entries(
 
     info!("user {user}: POST /datasets/{index_id}/get_entries",);
 
-    check_permission(&user, &index_id, Permission::Read, &findex_server).await?;
+    findex_server
+        .check_permission(&user, &index_id, Permission::Read)
+        .await?;
 
     let index_id = Uuid::parse_str(&index_id)?;
     let uuids = Uuids::deserialize(&bytes)?;
