+++ crate/server/src/routes/mod.rs
@@ -7,6 +7,6 @@ mod version;
 pub(crate) use datasets::{datasets_add_entries, datasets_del_entries, datasets_get_entries};
 pub(crate) use findex::{findex_batch_read, findex_guarded_write};
 pub(crate) use permissions::{
-    check_permission, create_index_id, grant_permission, list_permission, revoke_permission,
+    create_index_id, grant_permission, list_permission, revoke_permission,
 };
 pub(crate) use version::get_version;
